Add feature for coloring the segment.
Extend hooks status report by an optional color value.
A list of tuples, consisting of queries and color can be used.
The list is iterated through until no query gets a hit.
The color of the first hit will be used to highlight or none.
The results are a sub-set of the unread messages, cause the queries
extend the basic one.