Provide query customization in configurations.
It is possible to define the unread query as well as the color query
tuples inside the users configurations now.
The default configuration has been adjusted to get the queries form the
hook, which were hard-coded so far.
The segment script writes the queries to files on start-up now.
The Python hook tries to read and parse the query files.